[Http] Move UploadErrorExceptionMessage constant to more exposed namespace.
